requirements.txt
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
load-dotenv==0.1.0
pillow==10.4.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
pystray==0.19.5
python-dotenv==1.0.1
python-xlib==0.33
redis==5.0.8
requests==2.32.3
schedule==1.2.2
six==1.16.0
spotipy==2.24.0
urllib3==2.2.2
